requirements.txt
# API Clients
requests>=2.31.0
feedparser>=6.0.10
cloudinary>=1.36.0

# NewsAPI.ai (EventRegistry SDK)
eventregistry>=9.0

# Hugo content generation
python-frontmatter>=1.1.0
python-slugify>=8.0.1

# Utilities
python-dateutil>=2.8.2
tenacity>=8.2.3

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
responses>=0.24.0
